Add more files to organize code
- Add defless_ui to manage unfunctional widgets more effectively

- Move switch variables to switches file for common imports

- Create timer_control file to kick off update_timer function and end game when needed
