+++ switches.py
@@ -1,0 +1,6 @@
+from tkinter import *
+
+#These work like On/Off Switches: App Starts in Dark mode (lights_off) and Muted. 
+lights_off = True 
+muted_app = True
+timer_on = False
